moved logic to another file
